refactor(commands): extract undo/revert/checkpoint commands into history.rs
